cw-better-be-social/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
celery==5.0.5
certifi==2022.12.7
cffi==1.15.1
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==3.1.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==39.0.2
daphne==2.5.0
Django==3.0.3
django-bootstrap4==2.2.0
djangorestframework==3.11.1
factory-boy==3.2.1
Faker==17.6.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==0.6.2
oauthlib==3.2.2
Pillow==8.3.2
prompt-toolkit==3.0.38
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.3.1
redis==3.5.3
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==3.0.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
